requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
asn1crypto==1.5.1
attrs==24.2.0
Automat==24.8.1
beautifulsoup4==4.12.3
bitarray==3.0.0
blake2b-py==0.3.0
blinker==1.8.2
cached-property==1.5.2
cbor2==5.6.5
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
ckzg==1.0.2
click==8.1.7
coincurve==20.0.0
coloredlogs==15.0.1
configparser==7.1.0
constantly==23.10.4
contourpy==1.3.0
coverage==6.5.0
cryptography==43.0.3
crytic-compile==0.3.7
cssselect==1.2.0
cycler==0.12.1
cytoolz==1.0.0
defusedxml==0.7.1
dictionaries==0.0.2
distlib==0.3.9
eth-abi==4.2.1
eth-account==0.11.3
eth-bloom==1.0.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.2
eth-utils==2.3.2
ethereum_input_decoder==0.2.2
filelock==3.16.1
Flask==3.0.3
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
hexbytes==0.2.3
huggingface-hub==0.26.1
humanfriendly==10.0
hyperlink==21.0.0
identify==2.6.1
idna==3.10
importlib_resources==6.4.5
incremental==24.7.2
iniconfig==2.0.0
itemadapter==0.9.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lru-dict==1.2.0
lxml==5.3.0
MarkupSafe==2.0.1
matplotlib==3.9.2
mock==5.1.0
mpmath==1.3.0
multidict==6.1.0
mypy_extensions==0.4.4
mythril==0.24.8
networkx==3.4.2
nodeenv==1.9.1
numpy==2.1.2
packaging==24.1
parsel==1.9.1
parsimonious==0.9.0
persistent==6.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
pre-commit==2.20.0
prettytable==3.11.0
propcache==0.2.0
Protego==0.3.1
protobuf==5.28.3
psutil==6.1.0
py-ecc==1.4.7
py-evm==0.7.0a1
py-flags==1.1.4
py-solc==3.2.0
py-solc-x==1.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
PyDispatcher==2.0.7
pyethash==0.1.27
pyOpenSSL==24.2.1
pyparsing==2.4.7
pytest==8.3.3
pytest-cov==5.0.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
pyunormalize==16.0.0
PyYAML==6.0.2
queuelib==1.7.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
rlp==3.0.0
rpds-py==0.20.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
Scrapy==2.11.2
semantic-version==2.10.0
service-identity==24.2.0
setuptools==75.2.0
six==1.16.0
slither-analyzer==0.10.4
solc-select==1.0.4
sortedcontainers==2.4.0
soupsieve==2.6
style==1.1.0
sympy==1.13.1
threadpoolctl==3.5.0
tldextract==5.1.2
tokenizers==0.20.1
toml==0.10.2
toolz==1.0.0
torch==2.5.0
torch-geometric==2.6.1
tqdm==4.66.5
transaction==5.0
transformers==4.46.0
trie==2.2.0
Twisted==24.7.0
typing_extensions==4.12.2
update==0.0.1
urllib3==2.2.3
virtualenv==20.27.0
w3lib==2.2.1
wcwidth==0.2.13
web3==6.20.3
websockets==13.1
Werkzeug==3.0.6
yarl==1.16.0
z3-solver==4.12.5.0
zope.deferredimport==5.0
zope.interface==7.1.1
zope.proxy==6.1
